Move `_socket_timeout` into the base socket transport
